requirements.txt
numpy>=1.24,<2.0
Pillow>=10.0,<11.0
rasterio>=1.3,<2.0
scipy>=1.11,<2.0
pystac-client>=0.7,<1.0
planetary-computer>=1.0,<2.0
